chore: Update SIWE message verification error handling for clarity when it fails - don't hide inner failure reason - return 401 if SIWE is reason for error
